Finish grid line functionality
Started out with a loop that changed the class name for each element
in the sketch grid. This caused lots of trouble with various selectors
used in the rest of my code.

Ended up going for a loop that directly changes the style attribute of
each element in the sketch grid.